Complete day 31 - Flash Card App
This is a simple flash card app that allows the user to test their French words memory. The user can interact with the program using the buttons.
In this program the user can select whether they already know the current word or not by simply clicking the right button or the wrong button.
If the user clicks the right button, the row containing that word and it's translation will be deleted from the words_to_learn.csv.
So after that the program would only show the words in words_to_learn.csv and it wouldn't show the words that the user already knows.
If the user runs the program for the first time or the words_to_learn.csv file is deleted, the program would consider all the words are unknown.